nyumbani/requirements.txt
django==3.2.8
django-oauth-toolkit==2.0.0
djangorestframework==3.11.2
django-cors-middleware==1.5.0
django-cors-headers==3.13.0
# psycopg2>=2.8,<2.9
gunicorn
uvicorn[standard]==0.17.6

dj-database-url==0.5.0
PyJWT==1.7.1
djangorestframework-simplejwt==4.5.0
django-extra-fields==3.0.2
django-filter==2.4.0
djoser==2.0.5
# pandas==1.1.2
xlrd==1.2.0
ipython
africastalking==1.2.3

drf-renderer-xlsx==0.3.9
jedi==0.17.2
Pillow==8.2.0

workalendar==15.4.0
holidays==0.11.2
psycopg2-binary

sendgrid==6.9.4
openpyxl==3.0.9

pyproj

channels==3.0.4
channels-redis==3.4.0

sphinx==5.0.1

django-utils-six==2.0

googlemaps==4.6.0

twilio==6.63.2
phonenumbers==8.12.51
django-phonenumber-field==6.3.0
django-ckeditor==6.4.2
bs4==0.0.1

requests==2.28.0
sentry-sdk==1.9.5
PyPDF2==2.11.1
djangorestframework-csv==2.1.1

supabase==1.0.3
realtime==1.0.0
slack-bolt==1.17.1
firebase-admin==6.1.0
django-forestadmin==1.4.11
